+++ src/main/java/calculator/parser/ExpressionValidator.java
@@ -1,0 +1,5 @@
+package calculator.parser;
+
+public interface ExpressionValidator {
+    void validate(String expression);
+}
